==> 01_read_data.ipynb <==
## Reading the data

import pandas as pd
import os.path

We will read the data describing the gene expression intesities - the data were acquired with DNA microarrays:

df = pd.read_csv(os.path.join('data','data.txt'), comment="#", sep="\t")

Which columns are available?

df.columns

The columns `'1 Avg (log2)'` and `'2 Avg (log2)'` contain averaged intensitites (obtained from three control replicates and three replicates in which ESRP2 was silenced). The last 6 columns contain intensities measured in all 6 samples (3 x control, 3 x knockdown). The experiment is describe in the [paper](https://www.mdpi.com/1422-0067/22/14/7477).

We will only focus on the columns with average intesities. We will also need gene names (symbols) also used in GEM:

df1 = df[['Gene Symbol', '1 Avg (log2)']]
df2 = df[['Gene Symbol', '2 Avg (log2)']] 

We will store these data in two separate files:

df1.columns=['gene', 'value']
df2.columns=['gene', 'value']

df1.to_csv(os.path.join('data','data_control.txt'), index=False)
df2.to_csv(os.path.join('data','data_kd.txt'), index=False)
